+++ internal/commands/kubernetes/list.go
@@ -20,16 +20,15 @@ type listCommand struct {
 	*commands.BaseCommand
 }
 
-// InitCommand implements Command.InitCommand
 func (s *listCommand) InitCommand() {
-	// Deprecating k8s
+	// Deprecating uks
 	// TODO: Remove this in the future
-	commands.SetSubcommandDeprecationHelp(s, []string{"k8s"})
+	commands.SetSubcommandDeprecationHelp(s, []string{"uks"})
 }
 
 // ExecuteWithoutArguments implements commands.NoArgumentCommand
 func (s *listCommand) ExecuteWithoutArguments(exec commands.Executor) (output.Output, error) {
-	// Deprecating k8s
+	// Deprecating uks
 	// TODO: Remove this in the future
 	commands.SetSubcommandExecutionDeprecationMessage(s, []string{"k8s"}, "uks")