+++ internal/commands/kubernetes/versions.go
@@ -18,16 +18,15 @@ type versionsCommand struct {
 	*commands.BaseCommand
 }
 
-// InitCommand implements Command.InitCommand
 func (s *versionsCommand) InitCommand() {
-	// Deprecating k8s
+	// Deprecating uks
 	// TODO: Remove this in the future
-	commands.SetSubcommandDeprecationHelp(s, []string{"k8s"})
+	commands.SetSubcommandDeprecationHelp(s, []string{"uks"})
 }
 
 // ExecuteWithoutArguments implements commands.NoArgumentCommand
 func (s *versionsCommand) ExecuteWithoutArguments(exec commands.Executor) (output.Output, error) {
-	// Deprecating k8s
+	// Deprecating uks
 	// TODO: Remove this in the future
 	commands.SetSubcommandExecutionDeprecationMessage(s, []string{"k8s"}, "uks")