+++ internal/commands/networkpeering/networkpeering.go
@@ -13,8 +13,11 @@ import (
 
 // BaseNetworkPeeringCommand creates the base "networkpeering" command
 func BaseNetworkPeeringCommand() commands.Command {
+	baseCmd := commands.New("network-peering", "Manage network peerings")
+	baseCmd.SetDeprecatedAliases([]string{"networkpeering"})
+
 	return &networkpeeringCommand{
-		commands.New("network-peering", "Manage network peerings"),
+		BaseCommand: baseCmd,
 	}
 }
 
@@ -27,7 +30,7 @@ func (np *networkpeeringCommand) InitCommand() {
 	np.Cobra().Aliases = []string{"np", "networkpeering"}
 	// Deprecating networkpeering in favour of network-peering
 	// TODO: Remove this in the future
-	commands.SetDeprecationHelp(np.Cobra(), []string{"networkpeering"})
+	commands.SetDeprecationHelp(np.Cobra(), np.DeprecatedAliases())
 }
 
 // waitForNetworkPeeringState waits for network peering to reach given state and updates progress message with key matching given msg. Finally, progress message is updated back to given msg and either done state or timeout warning.