+++ internal/commands/servergroup/servergroup.go
@@ -14,8 +14,11 @@ const (
 
 // BaseServergroupCommand creates the base "servergroup" command
 func BaseServergroupCommand() commands.Command {
+	baseCmd := commands.New("server-group", "Manage server groups")
+	baseCmd.SetDeprecatedAliases([]string{"servergroup"})
+
 	return &servergroupCommand{
-		commands.New("server-group", "Manage server groups"),
+		BaseCommand: baseCmd,
 	}
 }
 
@@ -28,7 +31,7 @@ func (sg *servergroupCommand) InitCommand() {
 	sg.Cobra().Aliases = []string{"sg", "servergroup"}
 	// Deprecating servergroup in favour of server-group
 	// TODO: Remove this in the future
-	commands.SetDeprecationHelp(sg.Cobra(), []string{"servergroup"})
+	commands.SetDeprecationHelp(sg.Cobra(), sg.DeprecatedAliases())
 }
 
 func stringsToServerUUIDSlice(exec commands.Executor, servers []string) (upcloud.ServerUUIDSlice, error) {