+++ CHANGELOG.md
@@ -13,6 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add termination_protection to upctl database show output
 - Experimental support for token authentication by defining token in `UPCLOUD_TOKEN` environment variable.
 - Experimental support for managing tokens with `account token` commands.
+- New command names and aliases added to improve consistency:
+    - Commands:
+        - load-balancer
+        - network-peering
+        - object-storage
+        - server-group
+    - Aliases:
+        - account: acc
+        - gateway: gw
+        - network-peering: np
+        - object-storage: obs
+        - partner: pr
+        - router: rt
+        - server: srv
+        - server-group: sg
+        - storage: st
 
 ### Changed
 
@@ -21,6 +37,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Prevent filename completion of flags that don't take filename args.
+
+### Deprecated
+- Deprecation of some commands and aliases ( new command names added to improve consistency )
+    - Deprecated commands: 
+        - loadbalancer
+        - networkpeering
+        - objectstorage
+        - servergroup
+    - Deprecated aliases:
+        - object-storage: objsto
+
 
 ## [3.14.0] - 2025-01-08